Add support for device misc data

+++ chemclipse/plugins/org.chromulan.system.control.devices/src/org/chromulan/system/control/devices/base/UlanControlDevice.java
@@ -110,7 +110,7 @@ public abstract class UlanControlDevice implements IUlanControlDevice {
 	}
 
 	@Override
-	public int getFlg() {
+	public long getFlg() {
 
 		return FLG_SUPPORT_CSD_CHROMATOGRAM;
 	}
