+++ chemclipse/plugins/org.chromulan.system.control/src/org/chromulan/system/control/device/IControlDevice.java
@@ -17,9 +17,9 @@ import org.chromulan.system.control.device.setting.IDeviceSetting;
 
 public interface IControlDevice extends Externalizable {
 
-	final int FLG_SUPPORT_CSD_CHROMATOGRAM = 0b1;
-	final int FLG_SUPPORT_MSD_CHROMATOGRAM = 0b10;
-	final int FLG_SUPPORT_WSD_CHROMATOGRAM = 0b100;
+	final long FLG_SUPPORT_CSD_CHROMATOGRAM = 0b1;
+	final long FLG_SUPPORT_MSD_CHROMATOGRAM = 0b10;
+	final long FLG_SUPPORT_WSD_CHROMATOGRAM = 0b100;
 
 	String getCompany();
 
@@ -29,7 +29,7 @@ public interface IControlDevice extends Externalizable {
 
 	IDeviceSetting getDeviceSetting();
 
-	int getFlg();
+	long getFlg();
 
 	String getModel();
